fix: prevent `author` and `publisher` deletion `if associated books exist` (#10)

+++ services/authors.service.js
@@ -1,6 +1,7 @@
 import createError from "http-errors";
 
 import authorRepository from "../repositories/author.repository.js";
+import booksService from "./books.service.js";
 
 /**
  * AuthorsService class.
@@ -58,13 +59,21 @@ class AuthorsService {
   }
 
   /**
-   * Deletes an author by their ID.
+   * Deletes an author by their ID if they have no books associated with them.
    * @param {string} authorID - The ID of the author to delete.
-   * @returns {Promise<void>} A promise that resolves when the author is successfully deleted.
-   * @throws {Error} If the author does not exist, has associated books, or if an error occurs while deleting the author.
+   * @returns {Promise<object>} A promise that resolves to the deleted author object.
+   * @throws {createError.NotFound} If the author does not exist.
+   * @throws {createError.Conflict} If the author has books associated with them and cannot be deleted.
+   * @throws {createError.InternalServerError} If an error occurs while deleting the author.
    */
   async deleteAuthor(authorID) {
     const findAuthor = await this.getAuthorByID(authorID);
+    const countBooksByAuthorID = await booksService.countBooksByAuthorID(findAuthor.id);
+
+    if (countBooksByAuthorID > 0) {
+      throw createError(409, "Author has books associated with them and cannot be deleted.");
+    }
+
     return await authorRepository.deleteAuthor(findAuthor.id);
   }
 }
